perf(cli): actually lazy-load command implementations
buildLazyCommand() provided no laziness: cli-executable.ts statically
imported { commandSpec } from every *.command.ts, and since each of those
modules held both the spec and the implementation, the static import pulled
in the whole module. esbuild then collapsed the dynamic import inside
buildLazyCommand entirely, so `varlock --version` eagerly parsed the full
proxy subsystem.

Split each command's spec into a sibling *.command-spec.ts holding only
name/description/args/examples, and register commands with gunshi's own
lazy(loader, definition) helper. The entry now imports specs only, and the
implementation arrives through a dynamic import when the command runs.

For the three commands with nested subcommands (proxy, cache, keychain),
the spec module declares each verb as its own lazy() whose loader pulls the
run fn out of the implementation module, so the parent spec no longer embeds
the implementations by construction.
