+++ .bumpy/lazy-command-loading.md
@@ -1,0 +1,5 @@
+---
+varlock: patch
+---
+
+lazy-load CLI command implementations so startup does not parse every command
