+++ packages/varlock/src/cli/commands/audit.command.ts
@@ -1,7 +1,6 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import ansis from 'ansis';
-import { define } from 'gunshi';
 
 import { FileBasedDataSource } from '../../env-graph';
 import { loadVarlockEnvGraph } from '../../lib/load-graph';
@@ -15,41 +14,9 @@ import {
 import { gracefulExit } from 'exit-hook';
 import { diffSchemaAndCodeKeys } from '../helpers/audit-diff';
 import { isWellKnownEnvKey } from '../helpers/well-known-env-keys';
+import { commandSpec } from './audit.command-spec';
 
-export const commandSpec = define({
-  name: 'audit',
-  description: 'Audit code env var usage against your .env.schema',
-  args: {
-    targets: {
-      type: 'positional',
-      required: false,
-      multiple: true,
-      description: 'Directories to scan for env var references (defaults to the current project)',
-    },
-    path: {
-      type: 'string',
-      short: 'p',
-      description: 'Path to a specific .env file or directory to use as the schema entry point',
-    },
-    ignore: {
-      type: 'string',
-      short: 'i',
-      multiple: true,
-      description: 'Directory to exclude from code scanning (can be specified multiple times)',
-    },
-  },
-  examples: `
-Scans your source code for environment variable references and compares them
-to keys defined in your varlock schema.
-
-Examples:
-  varlock audit                          # Audit current project
-  varlock audit --path .env.prod         # Audit using a specific env entry point
-  varlock audit ./src ./lib              # Only scan specific directories
-  varlock audit --ignore vendor          # Exclude a directory from scanning
-  varlock audit -i vendor -i generated   # Exclude multiple directories
-`.trim(),
-});
+export { commandSpec };
 
 function formatReference(cwd: string, ref: EnvVarReference): string {
   const relPath = path.relative(cwd, ref.filePath);
