+++ packages/varlock/src/cli/commands/doctor.command-spec.ts
@@ -1,0 +1,7 @@
+import { define } from 'gunshi';
+
+export const commandSpec = define({
+  name: 'doctor',
+  description: 'Debug and diagnose issues with your env file(s) and system',
+  args: {},
+});
