+++ packages/varlock/src/cli/commands/doctor.command.ts
@@ -1,12 +1,8 @@
-import { define } from 'gunshi';
 import { isBundledSEA } from '../helpers/install-detection';
 import { type TypedGunshiCommandFn } from '../helpers/gunshi-type-utils';
+import { commandSpec } from './doctor.command-spec';
 
-export const commandSpec = define({
-  name: 'doctor',
-  description: 'Debug and diagnose issues with your env file(s) and system',
-  args: {},
-});
+export { commandSpec };
 
 export const commandFn: TypedGunshiCommandFn<typeof commandSpec> = async (ctx) => {
   console.log('');
