+++ packages/varlock/src/cli/commands/encrypt.command.ts
@@ -1,4 +1,3 @@
-import { define } from 'gunshi';
 import { isCancel } from '@clack/prompts';
 import ansis from 'ansis';
 import path from 'node:path';
@@ -16,38 +15,9 @@ import { multiselect, password } from '../helpers/prompts';
 import { gracefulExit } from 'exit-hook';
 import * as localEncrypt from '../../lib/local-encrypt';
 import { writeBackValue } from '../../lib/local-encrypt/write-back';
+import { commandSpec } from './encrypt.command-spec';
 
-export const commandSpec = define({
-  name: 'encrypt',
-  description: 'Encrypt a value using device-local encryption',
-  args: {
-    'key-id': {
-      type: 'string',
-      description: 'Encryption key ID',
-      default: 'varlock-default',
-      // Hidden until multi-key round-trips: the varlock("local:...") reference does not
-      // encode a keyId, and the load-time resolver always decrypts with the default key,
-      // so encrypting with a non-default key produces values that cannot be loaded back.
-      hidden: true,
-    },
-    file: {
-      type: 'string',
-      description: 'Path to a .env file — encrypts all sensitive plaintext values in-place',
-    },
-  },
-  examples: `
-Encrypts a value using device-local encryption (Secure Enclave / TPM / file-based),
-producing a varlock("local:...") reference that is safe to commit.
-
-Single-value mode reads from stdin (or prompts interactively) so secrets stay out of
-shell history. --file mode encrypts all @sensitive plaintext values in a .env file in place.
-
-Examples:
-  echo "$MY_SECRET" | varlock encrypt    # Encrypt a value from stdin (non-interactive, agent-friendly)
-  varlock encrypt                        # Prompt interactively for a value
-  varlock encrypt --file .env.local      # Encrypt @sensitive plaintext values in a file in-place
-`.trim(),
-});
+export { commandSpec };
 
 async function encryptFile(keyId: string, filePath: string) {
   const resolvedPath = path.resolve(filePath);
