+++ packages/varlock/src/cli/commands/help.command-spec.ts
@@ -1,0 +1,7 @@
+import { define } from 'gunshi';
+
+export const commandSpec = define({
+  name: 'help',
+  description: 'Show help info for varlock',
+  args: {},
+});
