+++ packages/varlock/src/cli/commands/help.command.ts
@@ -1,11 +1,7 @@
-import { define } from 'gunshi';
 import { type TypedGunshiCommandFn } from '../helpers/gunshi-type-utils';
+import { commandSpec } from './help.command-spec';
 
-export const commandSpec = define({
-  name: 'help',
-  description: 'Show help info for varlock',
-  args: {},
-});
+export { commandSpec };
 
 export const commandFn: TypedGunshiCommandFn<typeof commandSpec> = async (ctx) => {
   // no-op - we'll trigger help from main entry point
