+++ packages/varlock/src/cli/commands/init.command.ts
@@ -3,7 +3,6 @@ import path from 'node:path';
 import fs from 'node:fs/promises';
 import ansis from 'ansis';
 import { isCancel, select } from '@clack/prompts';
-import { define } from 'gunshi';
 import { gracefulExit } from 'exit-hook';
 
 import { envSpecUpdater, parseEnvSpecDotEnvFile } from '@env-spec/parser';
@@ -24,33 +23,9 @@ import { tryCatch } from '@env-spec/utils/try-catch';
 import { scanCodeForEnvVars } from '../helpers/env-var-scanner';
 import { isWellKnownEnvKey } from '../helpers/well-known-env-keys';
 import { detectTypegenDecorator } from '../helpers/typegen-lang-detect';
+import { commandSpec } from './init.command-spec';
 
-export const commandSpec = define({
-  name: 'init',
-  description: 'Set up varlock in the current project',
-  args: {
-    agent: {
-      type: 'boolean',
-      description: 'Run in non-interactive mode for agent/automation workflows',
-    },
-  },
-  examples: `
-This command starts an interactive onboarding process to help you get started with Varlock.
-It will:
-  - Scan for existing .env files in your project
-  - Help create a .env.schema file from your .env.example or .env.sample file
-  - Install varlock as a dependency in package.json (if applicable)
-
-📍 Run this command in directories that contain .env or .env.* files
-
-Examples:
-  varlock init                    # Run in the current directory
-  varlock init --agent            # Run non-interactively (agent/automation friendly)
-  cd path/to/your/project && varlock init
-
-For more information, visit https://varlock.dev/getting-started/installation
-  `.trim(),
-});
+export { commandSpec };
 
 export const commandFn: TypedGunshiCommandFn<typeof commandSpec> = async (ctx) => {
   const agentMode = Boolean(ctx.values.agent);
