+++ packages/varlock/src/cli/commands/lock.command-spec.ts
@@ -1,0 +1,6 @@
+import { define } from 'gunshi';
+
+export const commandSpec = define({
+  name: 'lock',
+  description: 'Lock the encryption daemon, requiring biometric for next decrypt',
+});
