+++ packages/varlock/src/cli/commands/login.command-spec.ts
@@ -1,0 +1,7 @@
+import { define } from 'gunshi';
+
+export const commandSpec = define({
+  name: 'login',
+  description: 'Authenticate (using GitHub)',
+  args: {},
+});
